Android FCM: rename messaging-ios→native, fix Android device type, ignore android/ in eslint
- messaging-ios.ts renamed to messaging-native.ts: the @capacitor-firebase/messaging
  plugin works on both iOS and Android; old name was misleading
- useMessaging.ts: device type was hardcoded 'ios' for all native platforms.
  Added getNativeDeviceType() helper to return 'android' or 'ios' correctly.
  Updated all three device-type usages (createDeviceFromToken, disableMessaging,
  enabled check) and import reference.
- platform.ts: add isAndroidPlatform() helper
- eslint.config.js: add android/** to ignore list (build artifacts caused lint errors)
- CLAUDE.md: document google-services.json setup steps for Android FCM
- Android.md: update #300 status and steps

Remaining for #300: register Android app in Firebase console, download
google-services.json → packages/web/android/app/google-services.json
